+++ lightseq/csrc/layers_new/sdpa_layer.cpp
@@ -7,8 +7,8 @@ Scaled Dot Product Attention
 See paper "Attention is all you need" for details"
 */
 template <typename T1, typename T2>
-SDPALayer<T1, T2>::SDPALayer(int max_batch_tokens, int max_seq_len,
-                             int head_dim, int num_heads,
+SDPALayer<T1, T2>::SDPALayer(size_t max_batch_tokens, size_t max_seq_len,
+                             size_t head_dim, size_t num_heads,
                              float attn_prob_dropout_ratio)
     : Layer("SDPALayer"),
       // for training, max_batch_tokens =
